requirements.txt
PyQt6>=6.6.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
xlrd>=2.0.1
jinja2>=3.1.0
pyinstaller>=6.0.0
pillow>=10.0.0
